+++ bin/test.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# Configuration
-cd $(dirname "$0")/..
-
-echo "Volkswagen test"