+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="0.2.0"></a>
+# 0.2.0 (2018-02-22)
+
+
+### Docs
+
+* Update main readme.md ([a5e3985e40590682f7c275e2d09f175ce5feb715](https://github.com/red-panda-ci/testlenium/commit/a5e3985e40590682f7c275e2d09f175ce5feb715))
+
+### Update
+
+* First functional testlenium release ([d71ebea385633b4fb36bb3f8c43909d3b01f018b](https://github.com/red-panda-ci/testlenium/commit/d71ebea385633b4fb36bb3f8c43909d3b01f018b))
+* Remove testng publish reports ([586f292c73ac5e6de263a2594ca57f851863b880](https://github.com/red-panda-ci/testlenium/commit/586f292c73ac5e6de263a2594ca57f851863b880))
+
+
+
 <a name="0.1.0"></a>
 # 0.1.0 (2018-02-15)
 
@@ -5,6 +20,7 @@
 ### Build
 
 * Add CI stuff ([f1828622de9e7abd05bd8c9549b856dce985aecf](https://github.com/red-panda-ci/testlenium/commit/f1828622de9e7abd05bd8c9549b856dce985aecf))
+* Update CHANGELOG.md to v0.1.0 with Red Panda JPL ([b25dbe839d142a547a570b5d15f37c1bcf663c21](https://github.com/red-panda-ci/testlenium/commit/b25dbe839d142a547a570b5d15f37c1bcf663c21))
 
 ### Docs